pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.0",
    "flask>=3.1.1",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.0",
    "telethon>=1.40.0",
]
